+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/picking/process/PickingSlot_ChangeIsDynamic.java
@@ -22,30 +22,21 @@
 
 package de.metas.handlingunits.picking.process;
 
-import de.metas.handlingunits.model.I_M_PickingSlot;
 import de.metas.handlingunits.picking.PickingSlotService;
-import de.metas.handlingunits.picking.requests.ReleasePickingSlotRequest;
-import de.metas.i18n.AdMessageKey;
-import de.metas.picking.api.IPickingSlotDAO;
 import de.metas.picking.api.PickingSlotId;
 import de.metas.process.IProcessPrecondition;
 import de.metas.process.IProcessPreconditionsContext;
 import de.metas.process.JavaProcess;
 import de.metas.process.Param;
 import de.metas.process.ProcessPreconditionsResolution;
-import de.metas.util.Services;
 import lombok.NonNull;
-import org.adempiere.exceptions.AdempiereException;
 import org.compiere.SpringContextHolder;
 
 public class PickingSlot_ChangeIsDynamic extends JavaProcess implements IProcessPrecondition
 {
-	private final static AdMessageKey SLOT_CANNOT_BE_RELEASED = AdMessageKey.of("de.metas.handlingunits.picking.process.SLOT_CANNOT_BE_RELEASED");
-
 	@Param(parameterName = "IsDynamic")
 	private boolean isDynamic;
 
-	private final IPickingSlotDAO slotDAO = Services.get(IPickingSlotDAO.class);
 	private final SpringContextHolder.Lazy<PickingSlotService> pickingSlotServiceLazy = SpringContextHolder.lazyBean(PickingSlotService.class);
 
 	@Override
@@ -62,43 +53,7 @@ public class PickingSlot_ChangeIsDynamic extends JavaProcess implements IProcess
 	@Override
 	protected String doIt() throws Exception
 	{
-		final I_M_PickingSlot pickingSlot = slotDAO.getById(PickingSlotId.ofRepoId(getRecord_ID()), I_M_PickingSlot.class);
-		if (isDynamic == pickingSlot.isDynamic())
-		{
-			//nothing to do
-			return MSG_OK;
-		}
-
-		if (pickingSlot.isDynamic())
-		{
-			turnOffDynamicAllocation(pickingSlot);
-		}
-		else
-		{
-			turnOnDynamicAllocation(pickingSlot);
-		}
-
+		pickingSlotServiceLazy.get().switchDynamicAllocation(PickingSlotId.ofRepoId(getRecord_ID()), isDynamic);
 		return MSG_OK;
-	}
-
-	private void turnOnDynamicAllocation(@NonNull final I_M_PickingSlot pickingSlot)
-	{
-		pickingSlot.setIsDynamic(true);
-		slotDAO.save(pickingSlot);
-	}
-
-	private void turnOffDynamicAllocation(@NonNull final I_M_PickingSlot pickingSlot)
-	{
-		final PickingSlotId pickingSlotId = PickingSlotId.ofRepoId(pickingSlot.getM_PickingSlot_ID());
-
-		final boolean released = pickingSlotServiceLazy.get().releasePickingSlot(ReleasePickingSlotRequest.ofSlotId(pickingSlotId));
-		if (!released)
-		{
-			throw new AdempiereException(SLOT_CANNOT_BE_RELEASED)
-					.markAsUserValidationError();
-		}
-
-		pickingSlot.setIsDynamic(false);
-		slotDAO.save(pickingSlot);
 	}
 }
