+++ backend/de.metas.swat/de.metas.swat.base/src/main/java/de/metas/picking/api/IPickingSlotDAO.java
@@ -33,5 +33,5 @@ public interface IPickingSlotDAO extends ISingletonService
 
 	Set<PickingSlotIdAndCaption> retrievePickingSlotIdAndCaptions(@NonNull PickingSlotQuery query);
 
-	I_M_PickingSlot save(@NonNull I_M_PickingSlot slot);
+	void save(@NonNull I_M_PickingSlot slot);
 }
